-added: TrueTypeFont importer & renderer (which is a WIP), BinaryFileReader (byte swapping), TableManager (which is used to switch between modes [and should only be used to do that]), a readme text and a license

+++ RGBTABLE/src/Serial.cpp
@@ -34,7 +34,7 @@ Serial::Serial(uint8_t identifier, unsigned int BAUD)
 
   //We're not yet connected
   this->connected = false;
-  int currentPort = 0;
+  size_t currentPort = 0;
 
   //iterate over all available com ports until we are connected
   while (!this->connected && currentPort < Serial::AvailableComPorts.size())
